Refactor: extract the code health analysis tools into their own module
